% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cal_log_score}
\alias{cal_log_score}
\title{Title}
\usage{
cal_log_score(sim_object, obs, ph)
}
\arguments{
\item{sim_object:}{a sample of a call to Beta_forecast_p}

\item{obs:}{observed prediction target}

\item{ph:}{prediction horizon}
}
\value{
the log score
}
\description{
Title
}
