% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_results_data_frames.R
\name{create_results_data_frames}
\alias{create_results_data_frames}
\title{Title}
\usage{
create_results_data_frames(data, prediction_time_inds, S_mean, S_Precision)
}
\arguments{
\item{data}{data frame outputted from from `reformat_data`}

\item{S_mean}{the number of harmonics for the mean term}

\item{S_Precision}{the number of harmonics for the precision term}

\item{prediction_time_indsindicators}{of the values one wants to predict from}
}
\value{
empty data_set_results data frame
data_set_results: a data frame which has the summarized scores with columns
"model" (model statement), "prediction_horizon",
"prediction_target" (date of predicted value), "log_score",
"pt_pred" (point prediction), "AE" (absolute error),
"interval_pred_lb_95" (95% Prediction Interval lower bound),
"interval_pred_ub_95" (95% Prediction Interval upper bound),
"interval_pred_lb_50" (50% Prediction Interval lower bound),
"interval_pred_ub_50" (50% Prediction Interval upper bound),
"precision" (predicted precision),
"shape1" (shape parameter 1 in beta distribution as a function of mean and precision),
"shape2" (shape parameter 2 in beta distribution as a function of mean and precision),
"DS_score" (log(variance)+(AE^2)/variance), "vari" (variance of prediction)
}
\description{
Title
}
\examples{
\dontrun{
dat <- reformat_data(data = seasons_data_weekly_update,
county_name = "MA_Middlesex",
training_seasons = c("2015-2016" "2016-2017" "2017-2018"),
lags=1)
data_frames <- create_results_data_frames(data=dat,
prediction_time_inds = which(dat$train == FALSE),
S_mean=4,
S_Precision=4)
}
}
