% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_model.R
\name{fit_beta_model}
\alias{fit_beta_model}
\title{Title}
\usage{
fit_beta_model(S_mean, S_Precision, lags, data)
}
\arguments{
\item{S_mean}{the number of harmonics for the mean term}

\item{S_Precision}{the number of harmonics for the precision term}

\item{lags}{the number of lags}

\item{data}{a data frame outputted from `reformat_data`}
}
\value{
the results of a try (either fitted model or an error) of the model
using the above parameters
}
\description{
Title
}
\examples{
\dontrun{
dat <- reformat_data(data = seasons_data_weekly_update,
county_name = "MA_Middlesex",
training_seasons = c("2015-2016", "2016-2017", "2017-2018"),
lags=1)
model <- fit_beta_model(S_mean=4, S_Precision=4, lags=1,
data=dat[train==T,])
}



}
