% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_1_step.R
\name{forecast_1_step}
\alias{forecast_1_step}
\title{Title}
\usage{
forecast_1_step(object, p = 1, start_value, coefs_subset)
}
\arguments{
\item{object:}{output of a call to `fit_beta_model`}

\item{p:}{lag vale}

\item{start_value:}{the p lagged values before the time being predicted}

\item{coefs_subset:}{the sine and cosine terms for the value being predicted}
}
\value{
one possible one week ahead forecast
}
\description{
Title
}
\examples{

\dontrun{
"example coming once I check the start_value and coefs_subset work
with updated code"
}

}
