% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_k_steps.R
\name{forecast_k_steps}
\alias{forecast_k_steps}
\title{Title}
\usage{
forecast_k_steps(
  k,
  num_sims,
  test_data,
  raw_data,
  time_index,
  prediction_time_ind,
  S_mean,
  S_Precision,
  lags,
  county,
  training_seasons,
  covar
)
}
\arguments{
\item{k:}{integer of the number of steps to forecast}

\item{num_sims:}{the number of forecasted simulations}

\item{test_data:}{only the data that is being tested and not trained}

\item{raw_data:}{the raw data that has the necessary columns
to run through `reformat_data`}

\item{time_index:}{the time index (row number in test_data)}

\item{prediction_time_ind:}{row index in entire data frame corresponding to the time index}

\item{S_mean:}{the number of harmonics for the mean term}

\item{S_Precision:}{the number of harmonics for the precision term}

\item{lags:}{the number of lags}

\item{county:}{the county name}

\item{covar:}{the sine and cosine covariates}
}
\value{

}
\description{
Title
}
