% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_season_predictions.R
\name{forecast_season_predictions}
\alias{forecast_season_predictions}
\title{Title}
\usage{
forecast_season_predictions(
  data,
  county,
  training_seasons,
  parameter_data,
  threshold_data,
  num_sims = 1000,
  dates_to_forecast = "all"
)
}
\arguments{
\item{data:}{data frame of data that has:
- a column of county names called 'Adjusted_County' in our data set
- a column denoting the season called 'Season' in our data set
- a column of influenza emergency department visits called 'ili_rate' in our data set
- a date column called 'last_day_week' in our data set
- a column denoting the week in the calendar year called 'Week' in our data set}

\item{county:}{string of the county you want to forecast for}

\item{training_seasons:}{character string of the seasons to be included
in training set}

\item{parameter_data:}{data frame of the parameters $S_v$, $S_{\phi}$ and
$p$ for the counties}

\item{threshold_data:}{data frame that gives the 40th, 67.5th and 95th threshold}
}
\value{
a list of 3 containing the forecasts(called results), runtime, and
quantile_matrix
results: filled in data_set_results data frame from create_results_data_frames
runtime: specification of how long this function took to run
quantile_matrix: filled in quantile_matrix data frame from create_results_data_frames
}
\description{
Title
}
\examples{
\dontrun{
forecast <- forecast_season_predictions_loops(
data = seasons_data_weekly_update,
county = "MA_Middlesex",
training_seasons = c("2015-2016" "2016-2017" "2017-2018"),
parameter_data =  param_by_AIC_by_county)
}



}
