% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution_severity_categories.R
\name{get_distribution_severity_categories}
\alias{get_distribution_severity_categories}
\title{Title}
\usage{
get_distribution_severity_categories(values, county, threshold_data, beta_fit)
}
\arguments{
\item{values}{the sine, cosine, and lag values for the given target week}

\item{county}{the county being forecasted for}

\item{threshold_data}{data frame of county thresholds}

\item{beta_fit}{the model fit for the given forecast date}
}
\description{
Title
}
