% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_log_score}
\alias{get_log_score}
\title{Get Log Score}
\usage{
get_log_score(value, county, threshold_data, pred_thresholds)
}
\arguments{
\item{value}{the value you want the log score of (true value)}

\item{county}{county the forecast was made}

\item{threshold_data}{data frame of threshold cutoffs by county}

\item{pred_thresholds}{the predicted threshold proportions array in low, mod, high, vhigh order}
}
\description{
Get Log Score
}
