% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_horizons.R
\name{get_prediction_horizons}
\alias{get_prediction_horizons}
\title{Title}
\usage{
get_prediction_horizons(prediction_time_ind, data)
}
\arguments{
\item{prediction_time_ind}{an indicator of the prediction time 
(row index in the data frame being passed in)}

\item{data}{a data frame containing season_week outputted from `reformat_data`}
}
\value{
a range of horizons to predict for
}
\description{
Title
}
\examples{

\dontrun{
dat <- reformat_data(data = seasons_data_weekly_update,
county_name = "MA_Middlesex",
training_seasons = c("2015-2016" "2016-2017" "2017-2018"), 
lags=1)

prediction_horizons <- get_prediction_horizons(
prediction_time_ind = which(dat$train == FALSE)[15],
data = dat)
}


}
