% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_severity_categories.R
\name{get_severity_categories}
\alias{get_severity_categories}
\title{Title}
\usage{
get_severity_categories(values, county, threshold_data)
}
\arguments{
\item{values}{}

\item{county}{}

\item{threshold_data}{}
}
\description{
Title
}
