% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_data.R
\name{reformat_data}
\alias{reformat_data}
\title{Title}
\usage{
reformat_data(data, county_name, training_seasons, lags)
}
\arguments{
\item{data}{data frame of data that has:
- a column of county names called 'Adjusted_County' in our data set
- a column denoting the season called 'Season' in our data set
- a column of influenza emergency department visits called 'ili_rate' in our data set
- a date column called 'last_day_week' in our data set
- a column denoting the week in the calendar year called 'Week' in our data set}

\item{county_name}{the name of the county being forecasted}

\item{training_seasons}{the list of seasons to train on}

\item{lags}{the number of lags}
}
\value{
A filtered data frame only consisting of "county_name"'s data with additional
columns to the one passed in.
These consist of:
- "train": an indicator of if the data for that row is in the training seasons
- "new_ili": a shifted value (increase of 1e3) of the reported "ili_rate"
- "week_number": the number of weeks in a season
- "season_week": integer corresponding to what week that date is within the given season
- columns corresponding to the sine and cosine terms
of j*2*pi*omega where j=1,...,5 and omega is the corresponding week in the
season divided by total weeks. These columns are denoted sine_InPeriodj and cos_InPeriodj
- columns of the logit of the lagged values denoted "pk" for k=1,...,"lags".
}
\description{
Title
}
\examples{
\dontrun{
dat <- reformat_data(data = seasons_data_weekly_update,
county_name = "MA_Middlesex",
training_seasons = c("2015-2016" "2016-2017" "2017-2018"),
lags=1)
}

}
